refactor(script,cart,Project): ♻️ Modularize code and fix cart bugs
