+++ index.html
@@ -85,6 +85,6 @@
     </div>
     <!-- ends: product purchase model -->
     <footer>Direitos reservados &#x20AC;</footer>
-    <script type="module" src="js/script.js"></script>
+    <script type="module" src="./js/main.js"></script>
   </body>
 </html>
